fix: rename components and fix typos
