+++ apps/web/components/Feature/ProfileLayout/Layout1/Collections/index.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import CreateProfile1Layout from "@/components/Layout/CreateProfile1";
-import CollectionList from "../../Layout2/Collections/ColletionList";
+import CollectionList from "../../Layout2/Collections/CollectionList";
 import Hero from "../Hero";
 
 export default function Collections() {
