+++ apps/web/components/Feature/ProfileLayout/Layout2/Collections/index.tsx
@@ -2,7 +2,7 @@
 
 import CreateProfileLayout from "../../../../Layout/CreateProfile2";
 import Hero from "../Hero";
-import CollectionList from "./ColletionList";
+import CollectionList from "./CollectionList";
 
 export default function Collections() {
   return (
