+++ apps/web/components/Feature/ProfileLayout/Layout3/Collections/index.tsx
@@ -2,7 +2,7 @@
 
 import CreateProfile1Layout from "@/components/Layout/CreateProfile1";
 import Hero from "../Hero";
-import CollectionList from "../../Layout2/Collections/ColletionList";
+import CollectionList from "../../Layout2/Collections/CollectionList";
 
 export default function Collections() {
   return (
