+++ apps/web/locals/da.json
@@ -1131,7 +1131,7 @@
     "users": {
       "tabs": {
         "registrations": "Registreringer (11)",
-        "salest": "Salg (50)"
+        "sales": "Salg (50)"
       },
       "registrations": {
         "title": "E-mail tilmeldinger",
