+++ apps/web/utils/usersTabs.ts
@@ -1,6 +1,6 @@
 export const USER_TABS = [
   { key: "registrations", labelKey: "users.tabs.registrations" },
-  { key: "salest", labelKey: "users.tabs.salest" },
+  { key: "sales", labelKey: "users.tabs.sales" },
 ] as const;
 
 export type UserTabKey = (typeof USER_TABS)[number]["key"];
